More changes

+++ BatteryApp/app/src/main/java/com/example/batteryapp/MainActivity.java
@@ -3,10 +3,16 @@ package com.example.batteryapp;
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.app.ActivityManager;
+import android.app.AlarmManager;
+import android.app.PendingIntent;
+import android.content.BroadcastReceiver;
+import android.content.ComponentName;
 import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
+import android.content.ServiceConnection;
 import android.os.Bundle;
+import android.os.IBinder;
 import android.util.Log;
 import android.view.View;
 import android.widget.Button;
@@ -15,6 +21,7 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import java.io.BufferedReader;
+import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
@@ -22,27 +29,49 @@ import java.io.IOException;
 import java.io.InputStreamReader;
 import java.text.SimpleDateFormat;
 import java.util.Date;
+import java.util.Scanner;
+import java.util.UUID;
 
 public class MainActivity extends AppCompatActivity implements View.OnClickListener {
     private static TextView textSOC, textTemperature, textVoltage, textTechnology, textStatus, textHealth, textTime, textCpuUsage;
     private static Button btnReadCpuFreq, btnSave, btnLoad;
     private static ProgressBar barCPU;
 
-    private static TextView textWifi, textData, textBluetooth, textBrightness, textRAM, mEditText, textTemp;
+    private static TextView textWifi, textData, textBluetooth, textBrightness, textRAM, mEditText, textTemp, textUUID;
 
     private final Battery_Receiver myBroadcast  = new Battery_Receiver();
     private final CPU_Info         myCPUInfo    = new CPU_Info();
     private final Other_Stats      myOtherStats = new Other_Stats(this);
+    private static String          userID;
 
     Intent myServiceIntent;
     private MyService myService;
-
     /* Called when the activity is first created. */
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         System.out.println("Inside Create");
         setContentView(R.layout.activity_main);
+        initializeComponents();
+
+        /* Create a unique id per user, only for the 1st time */
+        File f = new File(getFilesDir() + "/UUID.txt");
+        if ( !f.exists() ) {
+            // The first time create the file to store the ID
+            userID = UUID.randomUUID().toString();
+            userID = userID.replace("-", "");
+            setTextUUID("Your Unique userID is: " + userID);
+            saveFile(userID,"UUID.txt");
+        } else {
+            try {
+                Scanner myReader = new Scanner(f);
+                while (myReader.hasNextLine()) { userID = myReader.nextLine(); }
+                myReader.close();
+            } catch (FileNotFoundException e) {
+                e.printStackTrace();
+            }
+            setTextUUID("Your Unique userID is: " + userID);
+        }
 
         myService = new MyService();
         myServiceIntent = new Intent(this, myService.getClass());
@@ -50,7 +79,6 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
             startService(myServiceIntent);
         }
 
-        initializeComponents();
         // registerBatteryLevelReceiver();
 
         myOtherStats.getNetworkConnectivity();
@@ -62,9 +90,6 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
         btnReadCpuFreq.setOnClickListener( this );
         btnSave.setOnClickListener( this );
         btnLoad.setOnClickListener( this );
-
-
-
 
     }
 
@@ -83,14 +108,17 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
     @Override
     public void onClick(View v) {
         if (v == btnReadCpuFreq){
-            setCPUbar( (int) myCPUInfo.readCpuFreqNow() );
+            setCPUbar( (int) myCPUInfo.readCpuPercentNow() );
         } else if (v == btnSave) {
-            saveFile();
+            SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd_HH-mm-ss");
+            String todayDate = dateFormat.format(new Date());
+
+            String FILE_NAME = todayDate + ".txt";
+            saveFile("This is a test file", FILE_NAME);
         } else if (v == btnLoad) {
             loadFile();
         }
     }
-
 
     @Override
     protected void onStart() {
@@ -106,7 +134,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
         btnReadCpuFreq.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                setCPUbar( (int) myCPUInfo.readCpuFreqNow() );
+                setCPUbar( (int) myCPUInfo.readCpuPercentNow() );
             }
         });
     }
@@ -130,6 +158,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
         super.onDestroy();
         System.out.println("Inside Destroy");
         unregisterBatteryLevelReceiver();
+
         stopService(myServiceIntent);
     }
 
@@ -150,13 +179,8 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
         }
     }
 
-    /* Create a file based on timestamp */
-    private void saveFile() {
-        SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd_HH-mm-ss");
-        String todayDate = dateFormat.format(new Date());
-
-        String FILE_NAME = todayDate + ".txt";
-        String text = "I' am a test file";
+    /* Create a file with name "FILE_NAME" that contains "text" */
+    private void saveFile(String text, String FILE_NAME) {
         FileOutputStream fos = null;
         try {
             fos = openFileOutput(FILE_NAME, MODE_PRIVATE);
@@ -186,29 +210,31 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
             System.out.println(files[i]);
         }
         for (String FILE_NAME : files) {
-            try {
-                fis = openFileInput(FILE_NAME);
-                InputStreamReader isr = new InputStreamReader(fis);
-                BufferedReader br = new BufferedReader(isr);
-                StringBuilder sb = new StringBuilder();
-                String text;
+            if ( !FILE_NAME.contains("UUID")) {
+                try {
+                    fis = openFileInput(FILE_NAME);
+                    InputStreamReader isr = new InputStreamReader(fis);
+                    BufferedReader br = new BufferedReader(isr);
+                    StringBuilder sb = new StringBuilder();
+                    String text;
 
-                while ((text = br.readLine()) != null) {
-                    sb.append(text).append("\n");
-                }
+                    while ((text = br.readLine()) != null) {
+                        sb.append(text).append("\n");
+                    }
 
-                mEditText.setText(sb.toString());
-                deleteFile(FILE_NAME);
-            } catch (FileNotFoundException e) {
-                e.printStackTrace();
-            } catch (IOException e) {
-                e.printStackTrace();
-            } finally {
-                if (fis != null) {
-                    try {
-                        fis.close();
-                    } catch (IOException e) {
-                        e.printStackTrace();
+                    mEditText.setText(sb.toString());
+                    deleteFile(FILE_NAME);
+                } catch (FileNotFoundException e) {
+                    e.printStackTrace();
+                } catch (IOException e) {
+                    e.printStackTrace();
+                } finally {
+                    if (fis != null) {
+                        try {
+                            fis.close();
+                        } catch (IOException e) {
+                            e.printStackTrace();
+                        }
                     }
                 }
             }
@@ -238,6 +264,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
         btnLoad         = (Button) findViewById(R.id.btnLoad);
         mEditText       = (TextView) findViewById(R.id.textView);
         textTemp        = (TextView) findViewById(R.id.textTemp);
+        textUUID        = (TextView) findViewById(R.id.UserID);
 
     }
 
@@ -290,6 +317,9 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
 
     public void setCPUUsage(int usage) { textCpuUsage.setText("CPU Usage Estimation: " + usage + "%"); }
 
+    public void setTextUUID(String text) {
+        textUUID.setText(text);
+    }
     public void setTempText(String text) {
         textTemp.setText(text);
     }
